+++ lib/modules/account/bloc/builder.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:universy/modules/account/inputUser.dart';
 import 'package:universy/modules/account/login.dart';
-import 'package:universy/modules/account/newPassword.dart';
 import 'package:universy/modules/account/recover.dart';
 import 'package:universy/modules/account/signup.dart';
 import 'package:universy/modules/account/verify.dart';
@@ -20,8 +19,6 @@ class AccountStateBuilder extends WidgetBuilderFactory<AccountState> {
       return RecoverPasswordWidget(user: state.user,);
     } else if (state is InputUserState) {
       return InputUserWidget();
-    } else if (state is SetNewPasswordState) {
-      return SetNewPasswordWidget(user: state.user);
     }
 
     return LogInWidget();
