+++ lib/widgets/formfield/text/validators.dart
@@ -6,8 +6,7 @@ abstract class TextFormFieldValidatorBuilder
 
 typedef ValidationFunction = bool Function(String value);
 
-class NotEmptyTextFormFieldValidatorBuilder
-    implements TextFormFieldValidatorBuilder {
+class NotEmptyTextFormFieldValidatorBuilder implements TextFormFieldValidatorBuilder {
   final String _emptyMessage;
 
   NotEmptyTextFormFieldValidatorBuilder(this._emptyMessage);
@@ -23,16 +22,13 @@ class NotEmptyTextFormFieldValidatorBuilder
   }
 }
 
-class PatternNotEmptyTextFormFieldValidatorBuilder
-    implements TextFormFieldValidatorBuilder {
+class PatternNotEmptyTextFormFieldValidatorBuilder implements TextFormFieldValidatorBuilder {
   final RegExp pattern;
   final String _patternMessage;
   final String _emptyMessage;
 
   PatternNotEmptyTextFormFieldValidatorBuilder(
-      {@required RegExp regExp,
-      @required String patternMessage,
-      @required String emptyMessage})
+      {@required RegExp regExp, @required String patternMessage, @required String emptyMessage})
       : this.pattern = regExp,
         this._patternMessage = patternMessage,
         this._emptyMessage = emptyMessage;
@@ -50,8 +46,7 @@ class PatternNotEmptyTextFormFieldValidatorBuilder
   }
 }
 
-class NotEmptyFunctionTextFormValidatorBuilder
-    implements TextFormFieldValidatorBuilder {
+class NotEmptyFunctionTextFormValidatorBuilder implements TextFormFieldValidatorBuilder {
   final ValidationFunction _validationFunction;
   final String _message;
   final String _emptyMessage;
@@ -77,8 +72,7 @@ class NotEmptyFunctionTextFormValidatorBuilder
   }
 }
 
-class NotEqualTextFormValidatorBuilder
-    implements TextFormFieldValidatorBuilder {
+class NotEqualTextFormValidatorBuilder implements TextFormFieldValidatorBuilder {
   final ValidationFunction _validationFunction;
   final String _message;
   final String _emptyMessage;
@@ -112,8 +106,7 @@ class NotEqualTextFormValidatorBuilder
   }
 }
 
-class NotEqualTextFormValidatorBuilderPassword
-    implements TextFormFieldValidatorBuilder {
+class NotEqualTextFormValidatorBuilderPassword implements TextFormFieldValidatorBuilder {
   final RegExp pattern;
   final String _patternMessage;
   final String _emptyMessage;
@@ -122,12 +115,12 @@ class NotEqualTextFormValidatorBuilderPassword
 
   NotEqualTextFormValidatorBuilderPassword(
       {@required ValidationFunction validationFunction,
-        @required String message,
-        @required RegExp regExp,
-        @required String emptyMessage,
-        @required String patternMessage,
-        @required String notEqualMessage,
-        @required TextEditingController controllerToComparate})
+      @required String message,
+      @required RegExp regExp,
+      @required String emptyMessage,
+      @required String patternMessage,
+      @required String notEqualMessage,
+      @required TextEditingController controllerToComparate})
       : this._emptyMessage = emptyMessage,
         this._notEqualMessage = notEqualMessage,
         this.pattern = regExp,
@@ -139,10 +132,33 @@ class NotEqualTextFormValidatorBuilderPassword
     return (value) {
       if (value.isEmpty) {
         return _emptyMessage;
-      }else if (value.trim() != _controllerToComparate.text.trim()) {
+      } else if (value.trim() != _controllerToComparate.text.trim()) {
         return _notEqualMessage;
-      }else if (!pattern.hasMatch(value)) {
+      } else if (!pattern.hasMatch(value)) {
         return _patternMessage;
+      }
+      return null; // This null is required by the interface.
+    };
+  }
+}
+
+class NotEmptySixCharactersCodeValidator implements TextFormFieldValidatorBuilder {
+  final String _notQuantityValid;
+  final String _emptyMessage;
+
+  NotEmptySixCharactersCodeValidator({
+    @required String emptyMessage,
+    @required String notQuantityValid,
+  })  : this._notQuantityValid = notQuantityValid,
+        this._emptyMessage = emptyMessage;
+
+  @override
+  FormFieldValidator<String> build(BuildContext context) {
+    return (value) {
+      if (value.trim().length != 6) {
+        return _notQuantityValid;
+      } else if (value.isEmpty) {
+        return _emptyMessage;
       }
       return null; // This null is required by the interface.
     };
